requirements.txt
absl-py==2.3.1
accelerate==1.1.0
certifi==2025.8.3
charset-normalizer==3.4.3
contourpy==1.3.3
cycler==0.12.1
diffusers==0.32.2
einops==0.8.1
filelock==3.13.1
fonttools==4.60.0
fsspec==2024.6.1
grpcio==1.75.0
hf-xet==1.1.10
huggingface-hub==0.35.0
idna==3.10
importlib_metadata==8.7.0
Jinja2==3.1.4
joblib==1.5.2
kiwisolver==1.4.9
lpips==0.1.4
Markdown==3.9
MarkupSafe==2.1.5
matplotlib==3.10.6
mpmath==1.3.0
networkx==3.3
numpy==2.1.2
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.21.5
nvidia-nvtx-cu11==11.8.86
opencv-python==4.12.0.88
packaging==25.0
pillow==11.0.0
piq==0.8.0
protobuf==6.32.1
psutil==7.1.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
PyYAML==6.0.2
regex==2025.9.18
requests==2.32.5
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
setuptools==78.1.1
six==1.17.0
sympy==1.13.1
tensorboard==2.20.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
torch==2.6.0+cu118
torchaudio==2.6.0+cu118
torchvision==0.21.0+cu118
tqdm==4.67.1
triton==3.2.0
typing_extensions==4.12.2
urllib3==2.5.0
Werkzeug==3.1.3
wheel==0.45.1
zipp==3.23.0
xformers==0.0.29.post3
